refactor(time-entry): code moved to time-entry repository
