+++ app/Http/Controllers/TimeEntryController.php
@@ -8,7 +8,6 @@ use App\Models\TimeEntry;
 use App\Queries\TimeEntryDataTable;
 use App\Repositories\TimeEntryRepository;
 use Auth;
-use Carbon\Carbon;
 use DataTables;
 use Exception;
 use Illuminate\Contracts\View\Factory;
@@ -19,7 +18,6 @@ use Illuminate\Http\Response;
 use Illuminate\Validation\UnauthorizedException;
 use Illuminate\View\View;
 use Log;
-use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
 class TimeEntryController extends AppBaseController
 {
@@ -70,10 +68,7 @@ class TimeEntryController extends AppBaseController
      */
     public function store(CreateTimeEntryRequest $request)
     {
-        $input = $this->validateInput($request->all());
-
-        $this->timeEntryRepository->assignTaskToAdmin($input);
-        $this->timeEntryRepository->create($input);
+        $this->timeEntryRepository->store($request->all());
         $this->timeEntryRepository->broadcastStopTimerEvent();
 
         return $this->sendSuccess('Time Entry created successfully.');
@@ -110,7 +105,7 @@ class TimeEntryController extends AppBaseController
         if (empty($timeEntry)) {
             return $this->sendError('Time Entry not found.', Response::HTTP_NOT_FOUND);
         }
-        $input = $this->validateInput($request->all(), $timeEntry->id);
+
         $existEntry = $timeEntry->only([
             'id',
             'task_id',
@@ -121,6 +116,8 @@ class TimeEntryController extends AppBaseController
             'duration',
             'note',
         ]);
+
+        $input = $request->all();
         $inputDiff = array_diff($existEntry, $input);
         if (!empty($inputDiff)) {
             Log::info('Entry Id: '.$timeEntry->id);
@@ -150,48 +147,6 @@ class TimeEntryController extends AppBaseController
         $timeEntry->delete();
 
         return $this->sendSuccess('TimeEntry deleted successfully.');
-    }
-
-    /**
-     * @param array $input
-     * @param null  $id
-     *
-     * @return array|JsonResponse
-     */
-    public function validateInput($input, $id = null)
-    {
-        $startTime = Carbon::parse($input['start_time']);
-        $endTime = Carbon::parse($input['end_time']);
-        $input['duration'] = $endTime->diffInMinutes($startTime);
-        if ($startTime > $endTime) {
-            throw new BadRequestHttpException('Invalid start time and end time.');
-        }
-
-        $now = Carbon::now()->format('Y-m-d');
-        if ($startTime->format('Y-m-d') > $now) {
-            throw new BadRequestHttpException('Start time must be less than or equal to current time.');
-        }
-
-        if ($endTime->format('Y-m-d') > $now) {
-            throw new BadRequestHttpException('End time must be less than or equal to current time.');
-        }
-
-        if ($input['duration'] > 720) {
-            throw new BadRequestHttpException('Time Entry must be less than 12 hours.');
-        }
-
-        if ($input['duration'] < 1) {
-            throw new BadRequestHttpException('Minimum Entry time should be 1 minute.');
-        }
-
-        $this->timeEntryRepository->checkDuplicateEntry($input, $id);
-
-        $input['user_id'] = getLoggedInUserId();
-        if (!isset($input['note']) || empty($input['note'])) {
-            $input['note'] = 'N/A';
-        }
-
-        return $input;
     }
 
     /**
